feat: post 전체 리스트 페이지네이션 기능 추가 [#1]

+++ src/main/kotlin/com/kotlin/study/dongambackend/domain/post/repository/PostRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PostRepository : JpaRepository<Post, UInt> {
+interface PostRepository : JpaRepository<Post, Long> {
 
 }
